Implements sinatra-partial for mild refactoring opportunity
