+++ app/controllers/users.rb
@@ -1,0 +1,41 @@
+
+class Chitter < Sinatra::Base
+  get '/signup' do
+    @user = User.new
+    erb(:signup)
+  end
+
+  post '/signup' do
+    @user = User.create(name: params[:name], username: params[:username],
+                       email: params[:email], password: params[:password],
+                       password_confirmation: params[:confirm_password])
+    if @user.save
+      session[:user_id] = @user.id
+      redirect '/home'
+    else
+      flash.now[:signup_errors] = @user.errors.full_messages
+      erb(:signup)
+    end
+  end
+  
+  get '/login' do
+    erb(:login)
+  end
+
+  post '/login' do
+    @user = User.authenticate(params[:email], params[:password])
+    if @user
+      session[:user_id] = @user.id
+      redirect '/home'
+    else
+      flash.now[:login_error] = 'Invalid login details'
+      erb(:login)
+    end
+  end
+
+  delete '/logout' do
+    session[:user_id] = nil
+    flash.keep[:notice] = 'You logged out!'
+    redirect '/home'
+  end
+end
